Move common sass tools into separate file.
