+++ src/NodeTools/jsconfig.json
@@ -4,8 +4,5 @@
     },
     "exclude": [
         "node_modules"
-    ],
-    "typeAcquisition": {
-        "node": true
-    }
+    ]
 }
